% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendData.R
\name{AppendData}
\alias{AppendData}
\alias{AppendData.character}
\alias{AppendData.H5File}
\alias{AppendData.h5Seurat}
\title{Append data from an h5Seurat file to a preexisting
\code{\link[Seurat]{Seurat}} object}
\usage{
AppendData(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{character}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{H5File}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{AppendData}{h5Seurat}(
  file,
  object,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  extras = "commands",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file}{Name of h5Seurat or connected h5Seurat file to load}

\item{object}{A \code{\link[Seurat]{Seurat}} object to append data to}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
  \item \code{FALSE} for no assays
}}

\item{reductions}{One of:
\itemize{
 \item A character vector with names of reductions
 \item \code{NULL} for all reductions
 \item \code{NA} for \link[Seurat:IsGlobal]{global} reductions
 \item \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in
\code{assays} or marked as \link[Seurat:IsGlobal]{global} will be loaded}

\item{graphs}{One of:
\itemize{
 \item A character vector with names of graphs
 \item \code{NULL} for all graphs
 \item \code{FALSE} for no graphs
}
\strong{Note}: Only graphs associated with an assay loaded in \code{assays}
will be loaded}

\item{images}{One of:
\itemize{
 \item A character vector with names of images
 \item \code{NULL} for all images
 \item \code{NA} for \link[Seurat:IsGlobal]{global} images
 \item \code{FALSE} for no images
}}

\item{extras}{Extra information to load; supports any combination of the
following values:
\describe{
 \item{"commands"}{Load command logs. if \code{overwrite = TRUE}, replaces
 existing command logs}
}}

\item{overwrite}{Overwrite existing data in \code{object} with data from
\code{file}}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
\code{object} with the extra data requested
}
\description{
Append data from an h5Seurat file to a preexisting
\code{\link[Seurat]{Seurat}} object
}
