% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/loom.R
\name{as.loom}
\alias{as.loom}
\alias{as.loom.Seurat}
\title{Save an object as a loom file}
\usage{
as.loom(x, filename, overwrite = FALSE, verbose = TRUE, ...)

\method{as.loom}{Seurat}(x, filename, overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{filename}{Name of file to save \code{x} to}

\item{overwrite}{Overwrite \code{filename} if present?}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
A \code{\link{loom}} object
}
\description{
Save an object as a loom file
}
