% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteH5Group.R
\name{BasicWrite}
\alias{BasicWrite}
\title{Write data to an HDF5 dataset}
\usage{
BasicWrite(x, name, hgroup, verbose = TRUE)
}
\arguments{
\item{x}{An object}

\item{name}{Name to save data as}

\item{hgroup}{An HDF5 file or group (\code{H5File} or \code{H5Group} objects
from hdf5r)}

\item{verbose}{Show progress updates}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Write data to an HDF5 dataset
}
\keyword{internal}
