% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BoolToInt}
\alias{BoolToInt}
\title{Convert a logical to an integer}
\usage{
BoolToInt(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
An integer vector
}
\description{
Convert a logical to an integer
}
\examples{
\donttest{
SeuratDisk:::BoolToInt(x = c(TRUE, FALSE, NA))
}

}
\keyword{internal}
