% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{CheckMatrix}
\alias{CheckMatrix}
\title{Check that a dataset is a proper loom matrix}
\usage{
CheckMatrix(lfile, name, dims = NULL)
}
\arguments{
\item{lfile}{A \code{\link{loom}} or \code{\link[hdf5r]{H5File}} object}

\item{name}{Name of matrix to check}

\item{dims}{If provided, ensure \code{lfile[[name]]} has these dimensions;
should be a two-dimensional numberic vector with ncells/ncol as the first
value and nfeature/nrow as the second}
}
\value{
If all checks pass succesfully, invisibly returns \code{name}
}
\description{
Check that a dataset is a proper loom matrix
}
\keyword{internal}
