% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ChunkPoints}
\alias{ChunkPoints}
\title{Generate chunk points}
\usage{
ChunkPoints(dsize, csize)
}
\arguments{
\item{dsize}{Size of data being chunked}

\item{csize}{Size of chunk; if \code{NA}, assumes single chunk}
}
\value{
A matrix where each row is a chunk, column 1 is start points, column
2 is end points
}
\description{
Generate chunk points
}
\examples{
\donttest{
SeuratDisk:::ChunkPoints(100, 3)
SeuratDisk:::ChunkPoints(100, NA)
}

}
