% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{Convert}
\alias{Convert}
\alias{Convert.character}
\alias{Convert.H5File}
\alias{Convert.h5Seurat}
\title{Convert an on-disk single-cell dataset to another format}
\usage{
Convert(source, dest, assay, overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{character}(source, dest, assay, overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{H5File}(source, dest, assay, overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{h5Seurat}(
  source,
  dest = "h5ad",
  assay = DefaultAssay(object = source),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{assay}{Converting from \code{\link{h5Seurat}}: name of assay to write
out; converting to \code{\link{h5Seurat}}: name to store assay data as}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
If \code{source} is a \code{character}, invisibly returns \code{dest};
otherwise, returns an \code{\link[hdf5r]{H5File}}, or filetype-specific
subclass of \code{H5File} (eg. \code{\link{h5Seurat}}), connection to
\code{dest}
}
\description{
Convert an on-disk single-cell dataset to another format
}
