% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5info.R
\name{Dims}
\alias{Dims}
\title{Get the dimensions of an HDF5 dataset or sparse matrix}
\usage{
Dims(x)
}
\arguments{
\item{x}{An HDF5 dataset or sparse matrix}
}
\value{
A vector with the dimensions of the dataset or sparse matrix. For
sparse matrices, if no dimensions are found in either the \dQuote{dims} or
\dQuote{shape} attribute, returns \code{c(NA_integer_, NA_integer_)}
}
\description{
Get the dimensions of an HDF5 dataset or sparse matrix
}
\seealso{
\code{\href{../doc/h5Seurat-spec.html}{vignette("h5Seurat-spec")}}
}
\keyword{internal}
