% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Enumerate}
\alias{Enumerate}
\title{Enumerate a list or vector}
\usage{
Enumerate(x)
}
\arguments{
\item{x}{A list or a vector}
}
\value{
A list of length \code{x} with the following named values:
\describe{
  \item{\code{name}}{The name of \code{x} at a given index}
  \item{\code{value}}{The value of \code{x} at the corresponding index}
}
}
\description{
Enumerate a list or vector
}
\note{
For any given index \code{i} in \code{x}, all attempts to use the name
of the value of \code{x} at \code{i} will be made. If there is no name
(eg. \code{nchar(x = names(x = x)[i]) == 0}), the index \code{i} will be used
in its stead
}
\examples{
\donttest{
# Enumerate will use names if possible
x <- list(x = 1:3, y = letters[1:3], z = c('g1', 'g2', 'g3'))
SeuratDisk:::Enumerate(x = x)

# If no (or missing) object names present, Enumerate will use the index
# number as the name
x <- unname(obj = x)
SeuratDisk:::Enumerate(x = x)
}

}
\keyword{internal}
