% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5Seurat.R
\name{FetchCellData}
\alias{FetchCellData}
\title{Access cellular data}
\usage{
FetchCellData(object, vars, cells = NULL, slot = "data")
}
\arguments{
\item{object}{h5Seurat object}
}
\value{
A data frame with cells as rows and cellular data as columns
}
\description{
Retrieves data (feature expression, PCA scores, metrics, etc.) for a set
of cells in a Seurat object
}
