% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FileType}
\alias{FileType}
\title{Determine a filetype based on its extension}
\usage{
FileType(file)
}
\arguments{
\item{file}{Name of file}
}
\value{
The extension, all lowercase
}
\description{
Determine a filetype based on its extension
}
\examples{
\donttest{
SeuratDisk:::FileType('pbmc3k.h5Seurat')
SeuratDisk:::FileType('h5ad')
}

}
\keyword{internal}
