% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FixFeatures}
\alias{FixFeatures}
\title{Fix Feature Names}
\usage{
FixFeatures(features)
}
\arguments{
\item{features}{A vector of feature names}
}
\value{
Fixed features
}
\description{
Fix Feature Names
}
\keyword{internal}
