% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GetClass}
\alias{GetClass}
\title{Get a class string with package information}
\usage{
GetClass(class, packages = "Seurat")
}
\arguments{
\item{class}{Class name}

\item{packages}{A vector of packages to exclude from resulting class
information}
}
\value{
A character vector with the class
}
\description{
S4 classes are useful in the context of their defining package (benefits of
stricter typing). In order to ensure class information is properly retained
in HDF5 files, S4 class names are written as \dQuote{package:classname} with
certain exceptions (eg. S4 classes defined by
\link[Seurat:Seurat-package]{Seurat})
}
\examples{
\donttest{
SeuratDisk:::GetClass('Seurat')
SeuratDisk:::GetClass('Matrix')
}

}
\keyword{internal}
