% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GetMargin}
\alias{GetMargin}
\title{Determine the margin to use for a dataset}
\usage{
GetMargin(dims, MARGIN = getOption(x = "SeuratDisk.chunking.MARGIN"))
}
\arguments{
\item{dims}{Dimensions of a dataset}

\item{MARGIN}{Either an integer value contained within
\code{1:length(x = dims)} or one of the possible values of
\code{\link[SeuratDisk]{"SeuratDisk.chunking.MARGIN"}}}
}
\value{
An integer value with the \code{MARGIN}
}
\description{
Determine the margin to use for a dataset
}
\examples{
\donttest{
SeuratDisk:::GetMargin(c(4, 10))
}

}
\keyword{internal}
