% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObject.R
\name{GetObject}
\alias{GetObject}
\alias{GetAssays}
\alias{GetCommands}
\alias{GetDimReducs}
\alias{GetGraphs}
\alias{GetImages}
\title{Figure out which objects to load from an h5Seurat file}
\usage{
GetAssays(assays, index)

GetCommands(index, assays = NULL)

GetDimReducs(reductions, index, assays = NULL)

GetGraphs(graphs, index, assays = NULL)

GetImages(images, index)
}
\arguments{
\item{assays}{One of:
\itemize{
 \item A character vector with names of assays
 \item A character vector with one or more of \code{counts}, \code{data},
 \code{scale.data} describing which slots of \strong{all assays} to load
 \item A named list where each entry is either the name of an assay or a vector
 describing which slots (described above) to take from which assay
 \item \code{NULL} for all assays
}}

\item{index}{An h5Seurat index (\code{\link{h5SI}}) object}

\item{reductions}{One of:
\itemize{
 \item A character vector with names of reductions
 \item \code{NULL} for all reductions
 \item \code{NA} for \link[Seurat:IsGlobal]{global} reductions
 \item \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in
\code{assays} or marked as \link[Seurat:IsGlobal]{global} will be loaded}

\item{graphs}{One of:
\itemize{
 \item A character vector with names of graphs
 \item \code{NULL} for all graphs
 \item \code{FALSE} for no graphs
}
\strong{Note}: Only graphs associated with an assay loaded in \code{assays}
will be loaded}

\item{images}{One of:
\itemize{
 \item A character vector with names of images
 \item \code{NULL} for all images
 \item \code{NA} for \link[Seurat:IsGlobal]{global} images
 \item \code{FALSE} for no images
}}
}
\value{
\code{GetAssays}: A named list where each entry is a vector
describing the slots of an assay to load and the names are the assays to load

\code{GetCommands}" A vector of command log names that are derived
from an assay in \code{assay}

\code{GetDimReducs}: A vector of reduction names that are derived
from an assay in \code{assays} or global dimensional reductions

\code{GetGraphs}: A vector of graph names that are derived from an
assay in \code{assays}

\code{GetImages}: A vector of image names
}
\description{
Figure out which objects to load from an h5Seurat file
}
\seealso{
\code{\link{LoadH5Seurat}}
}
\keyword{internal}
