% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GetParent}
\alias{GetParent}
\title{Get the parent of an HDF5 dataset or group}
\usage{
GetParent(x)
}
\arguments{
\item{x}{An HDF5 dataset or group}
}
\value{
An \code{\link[hdf5r]{H5File}} or \code{\link[hdf5r]{H5Group}} object
}
\description{
Get the parent of an HDF5 dataset or group
}
\keyword{internal}
