% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GuessDType}
\alias{GuessDType}
\title{Guess an HDF5 Datatype}
\usage{
GuessDType(x, stype = "utf8", ...)
}
\arguments{
\item{x}{The object for which to guess the HDF5 datatype or the dimension or the number of elements}

\item{stype}{Type of string encoding to use, choose from:
\describe{
 \item{utf8}{Variable-width, UTF-8}
 \item{ascii7}{Fixed-width (7 bits), ASCII}
}}

\item{...}{
  Arguments passed on to \code{\link[hdf5r:guess_dtype]{hdf5r::guess_dtype}}
  \describe{
    \item{\code{ds_dim}}{Can explicitly set the dimension of the dataset object. For \code{scalar}, this is one. Otherwise, this can be
used so that a multi-dimensional object can be represented so that some of its dimension are in the dataset, and some are inside an
\code{\link[hdf5r]{H5T_ARRAY}}}
    \item{\code{scalar}}{Should the datatype be created so that \code{x} can be represented as a scalar with that datatype? This is intended
to know if a vector/array should be represented as an \code{\link[hdf5r]{H5T_ARRAY}} or not.}
    \item{\code{string_len}}{If a string is in the R object, the length to which the corresponding HDF5 type should be set. If it is a
positive integer, the string is of that length. If it is \code{Inf}, it is variable length. If it is set to \code{estimate},
it is set to the length of the longest string in the \code{x}.}
  }}
}
\value{
An object of class \code{\link[hdf5r]{H5T}}
}
\description{
Wrapper around \code{\link[hdf5r:guess_dtype]{hdf5r::guess_dtype}}, allowing
for the customization of string types rather than defaulting to
variable-length ASCII-encoded strings
}
\examples{
\donttest{
# Characters can either be variable-width UTF8-encoded or
# fixed-width ASCII-encoded
SeuratDisk:::GuessDType(x = 'hello')
SeuratDisk:::GuessDType(x = 'hello', stype = 'ascii7')

# Data frames are a compound type; character columns follow the same rules
# as character vectors
df <- data.frame(x = c('g1', 'g2', 'g3'), y = 1, 2, 3, stringsAsFactors = FALSE)
SeuratDisk:::GuessDType(x = df)
SeuratDisk:::GuessDType(x = df, stype = 'ascii7')

# Logicals are turned into integers to ensure compatability with Python
# TRUE evaluates to 1, FALSE to 0, and NA to 2
SeuratDisk:::GuessDType(x = c(TRUE, FALSE, NA))
}

}
\seealso{
\code{\link[hdf5r]{guess_dtype}}
}
\keyword{internal}
