% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5info.R
\name{H5Exists}
\alias{H5Exists}
\alias{AttrExists}
\alias{Exists}
\title{Check to see if a dataset, group, or attribute exists in an HDF5 file,
group, or dataset}
\usage{
AttrExists(x, name)

Exists(x, name)
}
\arguments{
\item{x}{An HDF5 \link[hdf5r:H5File]{file} or \link[hdf5r:H5Group]{group};
for \code{AttrExists}, may also be a \link[hdf5r:H5D]{dataset}}

\item{name}{Name of dataset, group, or attribute to test for}
}
\value{
\code{TRUE} if \code{name} exists in \code{x}, otherwise \code{FALSE}
}
\description{
Check to see if a dataset, group, or attribute exists in an HDF5 file,
group, or dataset
}
\keyword{internal}
