% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5info.R
\name{H5Path}
\alias{H5Path}
\title{Create an HDF5 object path}
\usage{
H5Path(..., collapse = NULL)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}
}
\value{
A character vector with path ready for accessing data in an HDF5 file
or group
}
\description{
Create an HDF5 object path
}
\keyword{internal}
