% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Seurat.R
\docType{class}
\name{h5Seurat-class}
\alias{h5Seurat-class}
\alias{h5Seurat}
\title{A class for connections to h5Seurat files}
\format{An \code{\link[R6]{R6Class}} object}
\description{
A class for connections to h5Seurat files

A class for connections to h5Seurat files
}
\seealso{
\code{\link[hdf5r]{H5File}}
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{hdf5r::H5RefClass}} -> \code{\link[hdf5r:H5File]{hdf5r::H5File}} -> \code{\link[SeuratDisk:scdisk]{SeuratDisk::scdisk}} -> \code{h5Seurat}
}
