% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Seurat.R
\name{h5SI}
\alias{h5SI}
\alias{DefaultAssay.h5SI}
\alias{print.h5SI}
\title{Tools for handling h5Seurat indexes}
\usage{
\method{DefaultAssay}{h5SI}(object)

\method{print}{h5SI}(x, ...)
}
\arguments{
\item{x, object}{An h5Seurat index (\code{h5SI})}
}
\value{
\code{DefaultAssay}: the assay set as the default assay

\code{print}: Invisibly returns \code{x}
}
\description{
Tools for handling h5Seurat indexes
}
\seealso{
\code{\link[Seurat]{DefaultAssay}} \code{\link[base]{print}}
}
\keyword{internal}
