% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{IsMatrixEmpty}
\alias{IsMatrixEmpty}
\title{Check to see if a matrix is empty}
\usage{
IsMatrixEmpty(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
\code{TRUE} if the matrix is empty otherwise \code{FALSE}
}
\description{
Determine if a matrix is empty or not. A matrix is considered empty if it
satisfies one of the following conditions:
\itemize{
 \item The dimensions of the matrix are 0-by-0 (\code{all(dim(x) == 0)})
 \item The dimensions of the matrix are 1-by-1 (\code{all(dim(x) == 1)}) and
 the sole vlaue is \code{NA}
}
These two situations correspond to matrices generated with either
\code{new('matrix')} or \code{matrix()}
}
\examples{
\donttest{
SeuratDisk:::IsMatrixEmpty(new('matrix'))
SeuratDisk:::IsMatrixEmpty(matrix())
SeuratDisk:::IsMatrixEmpty(matrix(1:9, nrow = 3))
}

}
\keyword{internal}
