% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdisk.R
\name{IsSCDisk}
\alias{IsSCDisk}
\title{Does an R6 class inherit from scdisk}
\usage{
IsSCDisk(r6class)
}
\arguments{
\item{r6class}{An \link[R6::R6Class]{R6 class generator} or a character name
of an R6 class generator}
}
\value{
If \code{r6class} inherits from scdisk, returns \code{TRUE};
otherwise, returns \code{FALSE}
}
\description{
Does an R6 class inherit from scdisk
}
\examples{
\donttest{
SeuratDisk:::IsSCDisk("H5File")
SeuratDisk:::IsSCDisk("scdisk")
SeuratDisk:::IsSCDisk("h5Seurat")
}

}
\keyword{internal}
