% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadLoom.R
\name{LoadLoom}
\alias{LoadLoom}
\alias{LoadLoom.character}
\alias{LoadLoom.H5File}
\alias{LoadLoom.loom}
\alias{as.Seurat.loom}
\alias{as.Seurat}
\title{Load data from a loom file into a \code{\link[Seurat]{Seurat}} object}
\usage{
LoadLoom(file, ...)

\method{LoadLoom}{character}(file, ...)

\method{LoadLoom}{H5File}(file, ...)

\method{LoadLoom}{loom}(file, ...)

\method{as.Seurat}{loom}(x, ...)
}
\arguments{
\item{file}{Name of h5Seurat or connected h5Seurat file to load}

\item{...}{Arguments passed to other methods}

\item{x}{Name of h5Seurat or connected h5Seurat file to load}
}
\value{
A \code{\link[Seurat]{Seurat}} object
}
\description{
Load data from a loom file into a \code{\link[Seurat]{Seurat}} object
}
