% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_bindings.R
\name{loom-bindings}
\alias{loom-bindings}
\alias{DefaultAssay.loom}
\alias{DefaultAssay}
\alias{dim.loom}
\alias{dim}
\title{Seurat binding for loom files}
\usage{
\method{DefaultAssay}{loom}(object, ...)

\method{dim}{loom}(x)
}
\description{
Seurat binding for loom files
}
