% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{MakeSpace}
\alias{MakeSpace}
\title{Make a space}
\usage{
MakeSpace(n)
}
\arguments{
\item{n}{Length space should be}
}
\value{
A space (' ') of length \code{n}
}
\description{
Generate a blank space \code{n} characters long; useful for aligning text to
be printed to console
}
\examples{
\donttest{
SeuratDisk:::MakeSpace(n = 10)
cat('hello', SeuratDisk:::MakeSpace(n = 10), 'world\n', sep = '')
}

}
\keyword{internal}
