% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PadMatrix.R
\name{PadMatrix}
\alias{PadMatrix}
\alias{PadMatrix,H5D-method}
\alias{PadMatrix,H5Group-method}
\title{Pad a matrix}
\usage{
PadMatrix(src, dest, dname)

\S4method{PadMatrix}{H5D}(src, dest, dname)

\S4method{PadMatrix}{H5Group}(src, dest, dname)
}
\arguments{
\item{...}{...}
}
\value{
...
}
\description{
Pad a matrix
}
\keyword{internal}
