% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PadNames}
\alias{PadNames}
\title{Add names for unnamed or partially named objects}
\usage{
PadNames(x, prefix = "index")
}
\arguments{
\item{x}{An object that can be named}

\item{prefix}{A prefix to be added to each name}
}
\value{
\code{x} with unnamed values named
}
\description{
Add names for unnamed or partially named objects
}
\examples{
\donttest{
a <- list(1, b = 2, 3)
SeuratDisk:::PadNames(a)
}

}
\keyword{internal}
