% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PB}
\alias{PB}
\title{Create a progress bar}
\usage{
PB()
}
\value{
An object of class \code{\link[utils]{txtProgressBar}}
}
\description{
Progress bars are useful ways of getting updates on how close a task is to
completion. However, they can get in the way of RMarkdown documents with
lots of unnecesssary printing. \code{PB} is a convenience function that
creates progress bars with the following defaults
\itemize{
 \item \code{char = '='}
 \item \code{style = 3}
 \item \code{file = stderr()}
}
}
\examples{
\donttest{
pb <- SeuratDisk:::PB()
for (i in 1:10) {
  utils::setTxtProgressBar(pb, i / 10)
}
close(pb)
}
}
\seealso{
\code{\link[utils]{txtProgressBar}} \code{\link[base]{stderr}}
}
\keyword{internal}
