% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{RandomName}
\alias{RandomName}
\title{Generate a random string of characters}
\usage{
RandomName(length = 5L, ...)
}
\arguments{
\item{length}{Length (\code{\link[base]{nchar}}) of string to generate}

\item{...}{Extra parameters passed to \code{\link[base]{sample}}}
}
\value{
A random string of characters of length (\code{\link[base]{nchar}})
of \code{length}
}
\description{
Generate a random string of characters
}
\examples{
\donttest{
SeuratDisk:::RandomName()
}

}
\keyword{internal}
