% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5.R
\name{ReadH5}
\alias{ReadH5}
\alias{as.array.H5D}
\alias{as.array}
\alias{as.data.frame.H5D}
\alias{as.data.frame}
\alias{as.data.frame.H5Group}
\alias{as.factor,H5Group-method}
\alias{as.factor}
\alias{as.list,H5Group-method}
\alias{as.list}
\alias{as.logical.H5D}
\alias{as.logical}
\alias{as.matrix.H5D}
\alias{as.matrix}
\alias{as.matrix.H5Group}
\alias{as.sparse.H5Group}
\alias{as.sparse}
\title{Load data from an HDF5 File}
\usage{
\method{as.array}{H5D}(x, ...)

\method{as.data.frame}{H5D}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{H5Group}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.factor}{H5Group}(x)

\S4method{as.list}{H5Group}(x, ...)

\method{as.logical}{H5D}(x, ...)

\method{as.matrix}{H5D}(x, transpose = FALSE, ...)

\method{as.matrix}{H5Group}(x, ...)

\method{as.sparse}{H5Group}(x, ...)
}
\arguments{
\item{x}{An HDF5 dataset or group}

\item{...}{Arguments passed to other methods}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{transpose}{Transpose the data upon reading it in, used when writing
data in row-major order (eg. from C or Python)}
}
\value{
\code{as.array}: returns an \code{\link[base]{array}} with the data
from the HDF5 dataset

\code{as.data.frame}: returns a \code{\link[base]{data.frame}} with
the data from the HDF5 dataset or group

\code{as.factor}: returns a \code{\link[base]{factor}} with the data
from the HDF5 group

\code{as.list}: returns a \code{\link[base]{list}} with the data from
the HDF5 group

\code{as.logical}: returns a \code{\link[base]{logical}} with the
data from the HDF5 dataset

\code{as.matrix}, \code{H5D} method: returns a
\code{\link[base]{matrix}} with the data from the HDF5 dataset

\code{as.sparse}; \code{as.matrix}, \code{H5Group} method: returns a
\code{\link[Matrix]{sparseMatrix}} with the data from the HDF5 group
}
\description{
HDF5 allows storing data in an arbitrary fashion, which makes reading data
into memory a hassle. The methods here serve as convenience functions for
reading data stored in a certain format back into a certain R object. For
details regarding how data should be stored on disk, please see the
\href{../doc/h5Seurat-spec.html}{h5Seurat file specification}.
}
