% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Seurat.R
\name{SaveH5Seurat}
\alias{SaveH5Seurat}
\title{Save an object as an h5seurat file}
\usage{
SaveH5Seurat(
  object = object,
  filename = paste0(Project(object = object), ".h5seurat"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filename}{Name of file to save \code{x} to}

\item{overwrite}{Overwrite \code{filename} if present?}

\item{verbose}{Show progress updates}

\item{...}{Extra parameters passed to and from other methods}
}
\value{
None, closes connection to file
}
\description{
Save an object as an h5seurat file
}
