% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{SaveLoom}
\alias{SaveLoom}
\title{Save an object as a loom file}
\usage{
SaveLoom(
  object,
  filename = "object.loom",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filename}{Name of file to save \code{x} to}

\item{overwrite}{Overwrite \code{filename} if present?}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
invisibly returns the loom file name
}
\description{
Save an object as a loom file
}
