% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Scalar}
\alias{Scalar}
\title{Create a scalar space}
\usage{
Scalar()
}
\value{
An object of type \code{\link[hdf5r:H5S]{H5S}} denoting a scalar HDF5
space
}
\description{
Create a scalar space
}
\keyword{internal}
