% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratDisk-pacakge}
\alias{SeuratDisk-pacakge}
\alias{_PACKAGE}
\alias{SeuratDisk}
\title{SeuratDisk: Some Title Here}
\description{
Some description here
}
\section{Package options}{


SeuratDisk uses the following options to control behaviour, users can configure
these with \code{\link[base]{options}}:

\describe{
 \item{\code{SeuratDisk.loom.encoding}, \code{SeuratDisk.loom3.encoding}}{
 Character encoding for loom files. The \code{loom3} encoding (default UTF-8)
 is used for loom files >= v3.0.0; otherwise, the \code{loom} encoding
 (default ASCII) is used}
 \item{\code{SeuratDisk.loom.string_len}}{When the loom encoding is ASCII, set
 the character width (default 7L)}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojaveazure/seurat-disk}
  \item Report bugs at \url{https://github.com/mojaveazure/seurat-disk/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Authors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
}

}
