% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{StringType}
\alias{StringType}
\title{Generate an HDF5 string dtype}
\usage{
StringType(stype = c("utf8", "ascii7"))
}
\arguments{
\item{stype}{Type of string encoding to use, choose from:
\describe{
 \item{utf8}{Variable-width, UTF-8}
 \item{ascii7}{Fixed-width (7 bits), ASCII}
}}
}
\value{
An \code{\link[hdf5r]{H5T_STRING}} object
}
\description{
Generate an HDF5 string dtype
}
\keyword{internal}
