% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestH5.R
\name{TestH5}
\alias{TestH5}
\alias{IsDataFrame,H5D-method}
\alias{IsDataFrame,H5Group-method}
\alias{IsFactor,H5D-method}
\alias{IsFactor,H5Group-method}
\alias{IsList,H5Group-method}
\alias{IsMatrix,H5D-method}
\alias{IsMatrix,H5Group-method}
\title{Test HDF5 datasets and groups to see what kind of data they are}
\usage{
\S4method{IsDataFrame}{H5D}(x)

\S4method{IsDataFrame}{H5Group}(x)

\S4method{IsFactor}{H5D}(x)

\S4method{IsFactor}{H5Group}(x)

\S4method{IsList}{H5Group}(x)

\S4method{IsMatrix}{H5D}(x)

\S4method{IsMatrix}{H5Group}(x)
}
\arguments{
\item{x}{An HDF5 dataset or group}
}
\description{
Test HDF5 datasets and groups to see what kind of data they are
}
\seealso{
\code{\link{TestObject}}
}
\keyword{internal}
