% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestObject.R
\name{TestObject}
\alias{TestObject}
\alias{IsDataFrame}
\alias{IsFactor}
\alias{IsList}
\alias{IsLogical}
\alias{IsMatrix}
\title{Test an object's class}
\usage{
IsDataFrame(x)

IsFactor(x)

IsList(x)

IsLogical(x)

IsMatrix(x)
}
\description{
Generic functions for testing if an object satisfies class membership
}
\keyword{internal}
