% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdisk.R
\name{Timestamp}
\alias{Timestamp}
\alias{FormatTime}
\alias{TSFormats}
\title{Create and work with timestamps}
\usage{
FormatTime(time, locale = TRUE, tz = "UTC", format = TSFormats(type = "R"))

Timestamp(tz = "UTC", format = TSFormats(type = "R"))

TSFormats(type = c("R", "loom"))
}
\arguments{
\item{time}{A character timestamp}

\item{locale}{Change the timestamp of to the timezone of the locale as
determined by \code{\link[base]{Sys.timezone}}}

\item{tz}{A character string specifying the time zone to be used for
    the conversion.  System-specific (see \code{\link[base]{as.POSIXlt}}), but
    \code{""} is the current time zone, and \code{"GMT"} is UTC.
    Invalid values are most commonly treated as UTC, on some platforms with
    a warning.}

\item{format}{A character string.  The default for the \code{format}
    methods is
    \code{"\%Y-\%m-\%d \%H:\%M:\%S"} if any element has a time
    component which is not midnight, and \code{"\%Y-\%m-\%d"}
    otherwise.  If \code{\link[base]{options}("digits.secs")} is set, up to
    the specified number of digits will be printed for seconds.}

\item{type}{Type of format to get, currently supports the following formats:
\itemize{
 \item \dQuote{R}
 \item \dQuote{loom}
}
See \strong{Timestamp formats} below for more details}
}
\value{
\code{FormatTime}: A character with the formated timestamp

\code{Timestamp}:A character with the current time in the
specified format

\code{TSFormats}: Format for a specified type
}
\description{
Create and work with timestamps
}
\section{Timestamp formats}{

The following formats can be provided by \code{TSFormats}:
\subsection{R}{
 A 24-hour R-friendly format; stores date/time information as the following:
 \itemize{
  \item Four-digit year (eg. \dQuote{2020} for the year 2020)
  \item Two-digit month (eg. \dQuote{04} for the month of April)
  \item Two-digit date (eg. \dQuote{02} for the second day of the month)
  \item The letter \dQuote{T}
  \item 24-hour two-digit time (eg. \dQuote{15} for 3:00 PM)
  \item Two-digit minute (eg \dQuote{05} for five minutes past the hour)
  \item Two-digit second (eg. \dQuote{05} for five seconds into the minute)
  \item The letter \dQuote{Z}
 }
 This results in a timestamp format of \dQuote{YYYYMMDDTHHMMSS.SSSSSSZ} (eg.
 \dQuote{20200402T150505Z} for April 4th, 2020 at 3:05:05 PM). \strong{Note}:
 this is considered \dQuote{R-friendly} as it does \emph{not} use precise
 values for seconds
}
\subsection{loom}{
 The standard format for loom timestamps; stores date/time information as
 the following:
 \itemize{
  \item Four-digit year (eg. \dQuote{2020} for the year 2020)
  \item Two-digit month (eg. \dQuote{04} for the month of April)
  \item Two-digit date (eg. \dQuote{02} for the second day of the month)
  \item The letter \dQuote{T}
  \item 24-hour two-digit time (eg. \dQuote{15} for 3:00 PM)
  \item Two-digit minute (eg \dQuote{05} for five minutes past the hour)
  \item Seconds precise to the millionth (six digits after the decimal)
  \item The letter \dQuote{Z}
 }
 This results in a timestamp format of \dQuote{YYYYMMDDTHHMMSS.SSSSSSZ} (eg.
 \dQuote{20200402T150505Z} for April 4th, 2020 at 3:05:05 PM). \strong{Note}:
 this is \emph{not} considered \dQuote{R-friendly} as it contains precise
 values for seconds. To properly format this time for pretty-printing, please
 remember to strip the precise seconds
}
}

\examples{
\donttest{
# Get a timestamp format
SeuratDisk:::TSFormats()

# Create a timestamp
SeuratDisk:::Timestamp()

# Format a timestamp for easy viewing
time <- "20200804T214148Z"
SeuratDisk:::FormatTime(time)
}

}
\keyword{internal}
