% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{UpdateKey}
\alias{UpdateKey}
\title{Update a Seurat key}
\usage{
UpdateKey(key)
}
\arguments{
\item{key}{A key to validate and update}
}
\value{
\code{key}, updated if invalid
}
\description{
Attempts to validate a string to use as a Seurat key. Valid keys must match
the regular expression \code{^[[:alnum:]]+_$}; if \code{key} fails this
regular expression, an attempt to modify it to said key will be made by
removing all non-alphanumeric characters, collapsing the resulting vector,
and appending \dQuote{_}. If this stil fails, a random string of lowercase
characters will be generated, followed by \dQuote{_}, to be used as the key
}
\examples{
\donttest{
SeuratDisk:::UpdateKey("RNA_")
SeuratDisk:::UpdateKey("potato")
SeuratDisk:::UpdateKey("*@)")
}

}
\seealso{
\code{\link[Seurat]{Key}} \code{\link{RandomName}}
}
\keyword{internal}
