% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{UpdateSlots}
\alias{UpdateSlots}
\title{Update slots in an object}
\usage{
UpdateSlots(object)
}
\arguments{
\item{object}{An object to update}
}
\value{
\code{object} with the latest slot definitions
}
\description{
Update slots in an object
}
\keyword{internal}
