% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Writeable}
\alias{Writeable}
\title{Is an HDF5 file or group writeable}
\usage{
Writeable(x, error = TRUE)
}
\arguments{
\item{x}{An \code{\link[hdf5r]{H5File}} or \code{\link[hdf5r]{H5Group}}
object}

\item{error}{Throw an error}
}
\value{
...
}
\description{
Is an HDF5 file or group writeable
}
\keyword{internal}
