% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{WriteAttribute}
\alias{WriteAttribute}
\title{Write an attribute to an HDF5 file, group, or dataset}
\usage{
WriteAttribute(x, name, lfile, stype, transpose = TRUE, verbose = TRUE)
}
\arguments{
\item{name}{Name to store attribute as}

\item{h5}{An HDF5 \link[hdf5r:H5File]{file}, \link[hdf5r:H5Group]{group}, or
\link[hdf5r:H5D]{dataset}}

\item{robj}{An object to write out}

\item{dtype}{Data type of attribute}

\item{scalar}{Is this a scalar or simple (vectorized) attribute?}

\item{overwrite}{Overwrite the attribute if it already exists}

\item{...}{Extra paramters passed to \code{\link[hdf5r:H5S]{H5S$new}}}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Write an attribute to an HDF5 file, group, or dataset
}
\keyword{internal}
