% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{WriteMode}
\alias{WriteMode}
\title{Get the proper HDF5 connection mode for writing depending on overwrite status}
\usage{
WriteMode(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite a file}
}
\value{
\code{w} if \code{overwrite} else \code{w-}
}
\description{
Get the proper HDF5 connection mode for writing depending on overwrite status
}
\examples{
\donttest{
SeuratDisk:::WriteMode(TRUE)
SeuratDisk:::WriteMode(FALSE)
}

}
\keyword{internal}
