% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.loom}
\alias{as.loom}
\title{Save an object as a loom file}
\usage{
as.loom(x, filename, overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{filename}{Name of file to save \code{x} to}

\item{overwrite}{Overwrite \code{filename} if present?}

\item{verbose}{Show progress updates}

\item{...}{Extra parameters passed to and from other methods}
}
\value{
...
}
\description{
Save an object as a loom file
}
