% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ClosestVersion}
\alias{ClosestVersion}
\title{Find the closest version}
\usage{
ClosestVersion(query, targets, direction = c("min", "max"))
}
\arguments{
\item{query}{A query version (\code{character} or \code{numeric_version})}

\item{targets}{A vector of target versions (\code{character} or
\code{numeric_version})}

\item{direction}{Which way should we check for closest version? Choose from:
\describe{
 \item{min}{Closest version less than or equal to \code{query}}
 \item{max}{Closest version greater than or equal to \code{query}}
}}
}
\value{
The version from \code{targets} that is closest to \code{query} as a
\code{character} vector
}
\description{
Find the closest version
}
\keyword{internal}
