% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{Convert}
\alias{Convert}
\alias{Convert.character}
\alias{Convert.H5File}
\alias{Convert.h5Seurat}
\title{Convert an on-disk single-cell dataset to another format}
\usage{
Convert(source, dest, assay, overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{character}(source, dest, assay, overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{H5File}(source, dest, assay = "RNA", overwrite = FALSE, verbose = TRUE, ...)

\method{Convert}{h5Seurat}(
  source,
  dest = "h5ad",
  assay = DefaultAssay(object = source),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{assay}{Converting from \code{\link{h5Seurat}}: name of assay to write
out; converting to \code{\link{h5Seurat}}: name to store assay data as}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
If \code{source} is a \code{character}, invisibly returns
\code{dest}; otherwise, returns an \code{\link[hdf5r]{H5File}}, or
filetype-specific subclass of \code{H5File} (eg. \code{\link{h5Seurat}}),
connection to \code{dest}
}
\description{
HDF5-based single-cell datasets can be converted from one format to another
using minimal memory. Details about conversion formats implemented are
provided below
}
\section{h5Seurat to AnnData/H5AD}{

The h5Seurat to AnnData/H5AD conversion will try to automatically fill in
datasets based on data presence. Data presense is determined by the h5Seurat
index (\code{source$index()}). It works in the following manner:
\subsection{Assay data:}{
 \itemize{
  \item \code{X} will be filled with \code{scale.data} if \code{scale.data}
  is present; otherwise, it will be filled with \code{data}
  \item \code{var} will be filled with \code{meta.features} \strong{only} for
  the features present in \code{X}; for example, if \code{X} is filled with
  \code{scale.data}, then \code{var} will contain only features that have
  been scaled
  \item \code{raw.X} will be filled with \code{data} if \code{X} is filled
  with \code{scale.data}; otherwise, it will be filled with \code{counts}. If
  \code{counts} is not present, then \code{raw} will not be filled
  \item \code{raw.var} will be filled with \code{meta.features} with the
  features present in \code{raw.X}; if \code{raw.X} is not filled, then
  \code{raw.var} will not be filled
 }
}
\subsection{Cell-level metadata:}{
 Cell-level metadata is added to \code{obs}
}
\subsection{Dimensional reduction information}{
 Only dimensional reductions associated with \code{assay} or marked as
 \link[Seurat:IsGlobal]{global} will be transfered to the H5AD file. For
 every reduction \code{reduc}:
 \itemize{
  \item cell embeddings are placed in \code{obsm} and renamed to
  \code{X_reduc}
  \item feature loadings, if present, are placed in \code{varm} and renamed
  to either \code{PCs} if \code{reduc} is "pca" otherwise \code{reduc} in all
   caps
 }
 For example, if \code{reduc} is "ica", then cell embeddings will be "X_ica"
 in \code{obsm} and feature loaodings, if present, will be "ICA" in
 \code{varm}
}
\subsection{Nearest-neighbor graphs}{
 If a nearest-neighbor graph is associated with \code{assay}, it will be
 added to \code{uns/neighbors/distances}; if more than one graph is present,
 then \strong{only} the last graph according to the index will be added.
}
}

