% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{H5ADToH5Seurat}
\alias{H5ADToH5Seurat}
\title{Convert AnnData/H5AD files to h5Seurat files}
\usage{
H5ADToH5Seurat(source, dest, assay = "RNA", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{source}{Source dataset}

\item{dest}{Name of destination dataset}

\item{assay}{Converting from \code{\link{h5Seurat}}: name of assay to write
out; converting to \code{\link{h5Seurat}}: name to store assay data as}

\item{overwrite}{Overwrite existing \code{dest}}

\item{verbose}{Show progress updates}
}
\value{
Returns a handle to \code{dest} as an \code{\link{h5Seurat}} object
}
\description{
Convert AnnData/H5AD files to h5Seurat files
}
\section{AnnData/H5AD to h5Seurat}{

The AnnData/H5AD to h5Seurat conversio will try to automatically fill in
datasets based on data presence. It works in the following manner:
\subsection{Expression data:}{
 The expression matrices \code{counts}, \code{data}, and \code{scale.data}
 are filled by \code{/X} and \code{/raw/X} in the following manner:
 \itemize{
  \item \code{counts} will be filled with \code{/raw/X} if present;
  otherwise, it will be filled with \code{/X}
  \item \code{data} will be filled with \code{/raw/X} if \code{/raw/X} is
  present and \code{/X} is dense; otherwise, it will be filled with \code{/X}
  \item \code{scale.data} will be filled with \code{/X} if it dense;
  otherwise, it will be empty
 }
 TODO: note about meta.features
}
\subsection{Cell-level metadata:}{
Cell-level metadata is added to \code{meta.data}; the row names of the
metadata (as determined by the value of the "_index" attribute or the "index"
dataset, in that order) are added to the "cell.names" dataset instead. If the
\code{__categories} dataset is present, each dataset within
\code{__categories} will be stored as a factor group. Cell-level metadata
will be added as an HDF5 group unless factors are \strong{not} present and
\code{\link[SeuratDisk]{SeuratDisk.dtype.dataframe_as_group}} is \code{FALSE}
}
\subsection{Dimensional reduction information:}{
 TODO: this
}
\subsection{Nearest-neighbor graph:}{
If a nearest neighbor graph is present in \code{/uns/neighbors/distances}, it
will be added as a graph dataset in the h5Seurat file and associated with
\code{assay}; if a value is present in \code{/uns/neighbors/params/metric},
the name of the graph will be \code{assay_metric}, otherwise, it will be
\code{assay_anndata}}
}

\keyword{internal}
