% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5Seurat.R
\docType{class}
\name{h5Seurat-class}
\alias{h5Seurat-class}
\alias{h5Seurat}
\title{A class for connections to h5Seurat files}
\format{
An \code{\link[R6]{R6Class}} object
}
\description{
A class for connections to h5Seurat files

A class for connections to h5Seurat files
}
\seealso{
\code{\link[hdf5r]{H5File}}
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{hdf5r::H5RefClass}} -> \code{\link[hdf5r:H5File]{hdf5r::H5File}} -> \code{\link[SeuratDisk:scdisk]{SeuratDisk::scdisk}} -> \code{h5Seurat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-h5Seurat-index}{\code{h5Seurat$index()}}
\item \href{#method-h5Seurat-set.version}{\code{h5Seurat$set.version()}}
\item \href{#method-h5Seurat-version}{\code{h5Seurat$version()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-close'><code>hdf5r::H5RefClass$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-dec_ref'><code>hdf5r::H5RefClass$dec_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_file_id'><code>hdf5r::H5RefClass$get_file_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_obj_type'><code>hdf5r::H5RefClass$get_obj_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-get_ref'><code>hdf5r::H5RefClass$get_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-inc_ref'><code>hdf5r::H5RefClass$inc_ref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods"><a href='../../hdf5r/html/H5RefClass.html#method-H5RefClass-methods'><code>hdf5r::H5RefClass$methods()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_delete'><code>hdf5r::H5File$attr_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_delete_by_idx'><code>hdf5r::H5File$attr_delete_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_delete_by_name'><code>hdf5r::H5File$attr_delete_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_exists"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_exists'><code>hdf5r::H5File$attr_exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_exists_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_exists_by_name'><code>hdf5r::H5File$attr_exists_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_get_number"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_get_number'><code>hdf5r::H5File$attr_get_number()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_info_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_info_by_idx'><code>hdf5r::H5File$attr_info_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_info_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_info_by_name'><code>hdf5r::H5File$attr_info_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_name_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_name_by_idx'><code>hdf5r::H5File$attr_name_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_open'><code>hdf5r::H5File$attr_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_open_by_idx'><code>hdf5r::H5File$attr_open_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_open_by_name'><code>hdf5r::H5File$attr_open_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_rename"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_rename'><code>hdf5r::H5File$attr_rename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_rename_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-attr_rename_by_name'><code>hdf5r::H5File$attr_rename_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="close_all"><a href='../../hdf5r/html/H5File.html#method-H5File-close_all'><code>hdf5r::H5File$close_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="commit"><a href='../../hdf5r/html/H5File.html#method-H5File-commit'><code>hdf5r::H5File$commit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_attr"><a href='../../hdf5r/html/H5File.html#method-H5File-create_attr'><code>hdf5r::H5File$create_attr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_attr_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-create_attr_by_name'><code>hdf5r::H5File$create_attr_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_dataset"><a href='../../hdf5r/html/H5File.html#method-H5File-create_dataset'><code>hdf5r::H5File$create_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_group"><a href='../../hdf5r/html/H5File.html#method-H5File-create_group'><code>hdf5r::H5File$create_group()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_reference"><a href='../../hdf5r/html/H5File.html#method-H5File-create_reference'><code>hdf5r::H5File$create_reference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="exists"><a href='../../hdf5r/html/H5File.html#method-H5File-exists'><code>hdf5r::H5File$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="file_info"><a href='../../hdf5r/html/H5File.html#method-H5File-file_info'><code>hdf5r::H5File$file_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="flush"><a href='../../hdf5r/html/H5File.html#method-H5File-flush'><code>hdf5r::H5File$flush()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_filename"><a href='../../hdf5r/html/H5File.html#method-H5File-get_filename'><code>hdf5r::H5File$get_filename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_filesize"><a href='../../hdf5r/html/H5File.html#method-H5File-get_filesize'><code>hdf5r::H5File$get_filesize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_intent"><a href='../../hdf5r/html/H5File.html#method-H5File-get_intent'><code>hdf5r::H5File$get_intent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_count"><a href='../../hdf5r/html/H5File.html#method-H5File-get_obj_count'><code>hdf5r::H5File$get_obj_count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_ids"><a href='../../hdf5r/html/H5File.html#method-H5File-get_obj_ids'><code>hdf5r::H5File$get_obj_ids()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_name"><a href='../../hdf5r/html/H5File.html#method-H5File-get_obj_name'><code>hdf5r::H5File$get_obj_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info"><a href='../../hdf5r/html/H5File.html#method-H5File-group_info'><code>hdf5r::H5File$group_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-group_info_by_idx'><code>hdf5r::H5File$group_info_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-group_info_by_name'><code>hdf5r::H5File$group_info_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link"><a href='../../hdf5r/html/H5File.html#method-H5File-link'><code>hdf5r::H5File$link()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_copy_from"><a href='../../hdf5r/html/H5File.html#method-H5File-link_copy_from'><code>hdf5r::H5File$link_copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_copy_to"><a href='../../hdf5r/html/H5File.html#method-H5File-link_copy_to'><code>hdf5r::H5File$link_copy_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_external"><a href='../../hdf5r/html/H5File.html#method-H5File-link_create_external'><code>hdf5r::H5File$link_create_external()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_hard"><a href='../../hdf5r/html/H5File.html#method-H5File-link_create_hard'><code>hdf5r::H5File$link_create_hard()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_soft"><a href='../../hdf5r/html/H5File.html#method-H5File-link_create_soft'><code>hdf5r::H5File$link_create_soft()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_delete"><a href='../../hdf5r/html/H5File.html#method-H5File-link_delete'><code>hdf5r::H5File$link_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_delete_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-link_delete_by_idx'><code>hdf5r::H5File$link_delete_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_exists"><a href='../../hdf5r/html/H5File.html#method-H5File-link_exists'><code>hdf5r::H5File$link_exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_info"><a href='../../hdf5r/html/H5File.html#method-H5File-link_info'><code>hdf5r::H5File$link_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_info_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-link_info_by_idx'><code>hdf5r::H5File$link_info_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_move_from"><a href='../../hdf5r/html/H5File.html#method-H5File-link_move_from'><code>hdf5r::H5File$link_move_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_move_to"><a href='../../hdf5r/html/H5File.html#method-H5File-link_move_to'><code>hdf5r::H5File$link_move_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_name_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-link_name_by_idx'><code>hdf5r::H5File$link_name_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_value"><a href='../../hdf5r/html/H5File.html#method-H5File-link_value'><code>hdf5r::H5File$link_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_value_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-link_value_by_idx'><code>hdf5r::H5File$link_value_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="ls"><a href='../../hdf5r/html/H5File.html#method-H5File-ls'><code>hdf5r::H5File$ls()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="mount"><a href='../../hdf5r/html/H5File.html#method-H5File-mount'><code>hdf5r::H5File$mount()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_copy_from"><a href='../../hdf5r/html/H5File.html#method-H5File-obj_copy_from'><code>hdf5r::H5File$obj_copy_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_copy_to"><a href='../../hdf5r/html/H5File.html#method-H5File-obj_copy_to'><code>hdf5r::H5File$obj_copy_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info"><a href='../../hdf5r/html/H5File.html#method-H5File-obj_info'><code>hdf5r::H5File$obj_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-obj_info_by_idx'><code>hdf5r::H5File$obj_info_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info_by_name"><a href='../../hdf5r/html/H5File.html#method-H5File-obj_info_by_name'><code>hdf5r::H5File$obj_info_by_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="open"><a href='../../hdf5r/html/H5File.html#method-H5File-open'><code>hdf5r::H5File$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="open_by_idx"><a href='../../hdf5r/html/H5File.html#method-H5File-open_by_idx'><code>hdf5r::H5File$open_by_idx()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="path_valid"><a href='../../hdf5r/html/H5File.html#method-H5File-path_valid'><code>hdf5r::H5File$path_valid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="print"><a href='../../hdf5r/html/H5File.html#method-H5File-print'><code>hdf5r::H5File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="unmount"><a href='../../hdf5r/html/H5File.html#method-H5File-unmount'><code>hdf5r::H5File$unmount()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeuratDisk" data-topic="scdisk" data-id="chunk.points"><a href='../../SeuratDisk/html/scdisk.html#method-scdisk-chunk.points'><code>SeuratDisk::scdisk$chunk.points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeuratDisk" data-topic="scdisk" data-id="finalizer"><a href='../../SeuratDisk/html/scdisk.html#method-scdisk-finalizer'><code>SeuratDisk::scdisk$finalizer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeuratDisk" data-topic="scdisk" data-id="initialize"><a href='../../SeuratDisk/html/scdisk.html#method-scdisk-initialize'><code>SeuratDisk::scdisk$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeuratDisk" data-topic="scdisk" data-id="last.modified"><a href='../../SeuratDisk/html/scdisk.html#method-scdisk-last.modified'><code>SeuratDisk::scdisk$last.modified()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeuratDisk" data-topic="scdisk" data-id="timestamp"><a href='../../SeuratDisk/html/scdisk.html#method-scdisk-timestamp'><code>SeuratDisk::scdisk$timestamp()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h5Seurat-index"></a>}}
\if{latex}{\out{\hypertarget{method-h5Seurat-index}{}}}
\subsection{Method \code{index()}}{
Get the index for this h5Seurat file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{h5Seurat$index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h5Seurat-set.version"></a>}}
\if{latex}{\out{\hypertarget{method-h5Seurat-set.version}{}}}
\subsection{Method \code{set.version()}}{
Set the version attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{h5Seurat$set.version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{A version number matching the regex
\code{^\\d+(\\.\\d+){2}(\\.9\\d{3})?$}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h5Seurat-version"></a>}}
\if{latex}{\out{\hypertarget{method-h5Seurat-version}{}}}
\subsection{Method \code{version()}}{
Get the version attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{h5Seurat$version()}\if{html}{\out{</div>}}
}

}
}
