% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{IsDType}
\alias{IsDType}
\title{Check the datatype of an HDF5 dataset}
\usage{
IsDType(x, dtype)
}
\arguments{
\item{x}{An HDF5 dataset (object of type \code{\link[hdf5r]{H5D}})}

\item{dtype}{A character vector of HDF5 datatype names, must be present in
\code{\link[hdf5r]{h5types}}}
}
\value{
A logical
}
\description{
Check the datatype of an HDF5 dataset
}
\seealso{
\code{\link[hdf5r]{h5types}}
}
\keyword{internal}
