% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadH5Seurat.R
\name{LoadH5Seurat}
\alias{LoadH5Seurat}
\alias{LoadH5Seurat.character}
\alias{LoadH5Seurat.H5File}
\alias{LoadH5Seurat.h5Seurat}
\alias{as.Seurat.h5Seurat}
\alias{as.Seurat}
\title{Load a saved \code{Seurat} object from an h5Seurat file}
\usage{
LoadH5Seurat(file, ...)

\method{LoadH5Seurat}{character}(
  file,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  meta.data = TRUE,
  commands = TRUE,
  misc = is.null(x = assays),
  tools = is.null(x = assays),
  verbose = TRUE,
  ...
)

\method{LoadH5Seurat}{H5File}(
  file,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  meta.data = TRUE,
  commands = TRUE,
  misc = is.null(x = assays),
  tools = is.null(x = assays),
  verbose = TRUE,
  ...
)

\method{LoadH5Seurat}{h5Seurat}(
  file,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  meta.data = TRUE,
  commands = TRUE,
  misc = is.null(x = assays),
  tools = is.null(x = assays),
  verbose = TRUE,
  ...
)

\method{as.Seurat}{h5Seurat}(
  x,
  assays = NULL,
  reductions = NULL,
  graphs = NULL,
  images = NULL,
  meta.data = TRUE,
  commands = TRUE,
  misc = is.null(x = assays),
  tools = is.null(x = assays),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{file, x}{Name of h5Seurat or connected h5Seurat file to load}

\item{...}{Arguments passed to other methods}

\item{assays}{One of:
\itemize{
 \item A character vector with names of assays
 \item A character vector with one or more of \code{counts}, \code{data},
 \code{scale.data} describing which slots of \strong{all assays} to load
 \item A named list where each entry is either the name of an assay or a vector
 describing which slots (described above) to take from which assay
 \item \code{NULL} for all assays
}}

\item{reductions}{One of:
\itemize{
 \item A character vector with names of reductions
 \item \code{NULL} for all reductions
 \item \code{NA} for \link[Seurat:IsGlobal]{global} reductions
 \item \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in
\code{assays} or marked as \link[Seurat:IsGlobal]{global} will be loaded}

\item{graphs}{One of:
\itemize{
 \item A character vector with names of graphs
 \item \code{NULL} for all graphs
 \item \code{FALSE} for no graphs
}
\strong{Note}: Only graphs associated with an assay loaded in \code{assays}
will be loaded}

\item{images}{One of:
\itemize{
 \item A character vector with names of images
 \item \code{NULL} for all images
 \item \code{NA} for \link[Seurat:IsGlobal]{global} images
 \item \code{FALSE} for no images
}}

\item{meta.data}{Load object metadata}

\item{commands}{Load command information \cr
\strong{Note}: only commands associated with an assay loaded in
\code{assays} will be loaded}

\item{misc}{Load miscellaneous data}

\item{tools}{Load tool-specific information}

\item{verbose}{Show progress updates}
}
\value{
A \code{Seurat} object with the data requested
}
\description{
Load a saved \code{Seurat} object from an h5Seurat file
}
