% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_bindings.R
\name{loom-bindings}
\alias{loom-bindings}
\alias{DefaultAssay.loom}
\alias{DefaultAssay}
\title{Seurat binding for loom files}
\usage{
\method{DefaultAssay}{loom}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Seurat binding for loom files
}
