% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\docType{class}
\name{loom-class}
\alias{loom-class}
\alias{loom}
\title{A class for connections to loom files}
\format{An \code{\link[R6]{R6Class}} object}
\description{
A class for connections to loom files

A class for connections to loom files
}
\seealso{
\code{\link[hdf5r]{H5File}}
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{hdf5r::H5RefClass}} -> \code{\link[hdf5r:H5File]{hdf5r::H5File}} -> \code{\link[SeuratDisk:scdisk]{SeuratDisk::scdisk}} -> \code{loom}
}
