% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveH5Seurat.R
\name{SaveH5Seurat}
\alias{SaveH5Seurat}
\alias{as.h5Seurat}
\alias{SaveH5Seurat.default}
\alias{SaveH5Seurat.Seurat}
\alias{as.h5Seurat.default}
\alias{as.h5Seurat.H5File}
\alias{as.h5Seurat.Seurat}
\title{Save a \code{Seurat} object to an h5Seurat file}
\usage{
SaveH5Seurat(object, filename, overwrite = FALSE, verbose = TRUE, ...)

as.h5Seurat(x, ...)

\method{SaveH5Seurat}{default}(object, filename, overwrite = FALSE, verbose = TRUE, ...)

\method{SaveH5Seurat}{Seurat}(
  object,
  filename = paste0(Project(object = object), ".h5Seurat"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{as.h5Seurat}{default}(x, filename, overwrite = FALSE, verbose = TRUE, ...)

\method{as.h5Seurat}{H5File}(x, ...)

\method{as.h5Seurat}{Seurat}(
  x,
  filename = paste0(Project(object = x), ".h5seurat"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object, x}{An object}

\item{filename}{Name of file to save the object to}

\item{overwrite}{Overwrite \code{filename} if present}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
\code{SaveH5Seurat}: Invisbly returns \code{filename}

\code{as.h5Seurat}: An \code{\link{h5Seurat}} object
}
\description{
Save a \code{Seurat} object to an h5Seurat file
}
