% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_h5Seurat.R
\name{SaveH5Seurat}
\alias{SaveH5Seurat}
\title{Save an object as an h5seurat file}
\usage{
SaveH5Seurat(
  object,
  filename = "object.h5seurat",
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filename}{Name of file to save \code{object} to}

\item{overwrite}{Overwrite \code{filename} if present?}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
Invisibly returns h5seurat file name
}
\description{
Save an object as an h5seurat file
}
\seealso{
\code{\link{as.h5Seurat}}
}
