% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoom.R
\name{SaveLoom}
\alias{SaveLoom}
\alias{as.loom}
\alias{SaveLoom.default}
\alias{SaveLoom.Seurat}
\alias{as.loom.default}
\alias{as.loom.H5File}
\alias{as.loom.Seurat}
\title{Save a \code{\link[Seurat]{Seurat}} object to a loom file}
\usage{
SaveLoom(object, filename, overwrite = FALSE, verbose = TRUE, ...)

as.loom(x, ...)

\method{SaveLoom}{default}(object, filename, overwrite = FALSE, verbose = TRUE, ...)

\method{SaveLoom}{Seurat}(
  object,
  filename = paste0(Project(object = object), ".loom"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{as.loom}{default}(x, filename, overwrite = FALSE, verbose = TRUE)

\method{as.loom}{H5File}(x, ...)

\method{as.loom}{Seurat}(
  x,
  filename = paste0(Project(object = x), ".loom"),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{filename}{Name of file to save the object to}

\item{overwrite}{Overwrite \code{filename} if present}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}

\item{x}{An object}
}
\value{
\code{SaveLoom}: Invisibly returns \code{filename}

\code{as.loom}: A \code{\link{loom}} object
}
\description{
Save a \code{\link[Seurat]{Seurat}} object to a loom file
}
