% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdisk.R
\docType{class}
\name{scdisk-class}
\alias{scdisk-class}
\alias{scdisk}
\title{A disk-based object for single-cell analysis}
\format{
An \code{\link[R6]{R6Class}} object
}
\description{
A disk-based object for single-cell analysis

A disk-based object for single-cell analysis
}
\seealso{
\code{\link[hdf5r]{H5File}}
}
\section{Super classes}{
\code{\link[hdf5r:H5RefClass]{hdf5r::H5RefClass}} -> \code{\link[hdf5r:H5File]{hdf5r::H5File}} -> \code{scdisk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{scdisk$new()}}
\item \href{#method-finalizer}{\code{scdisk$finalizer()}}
\item \href{#method-chunk.points}{\code{scdisk$chunk.points()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="close">}\href{../../hdf5r/html/H5RefClass.html#method-close}{\code{hdf5r::H5RefClass$close()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="dec_ref">}\href{../../hdf5r/html/H5RefClass.html#method-dec_ref}{\code{hdf5r::H5RefClass$dec_ref()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_file_id">}\href{../../hdf5r/html/H5RefClass.html#method-get_file_id}{\code{hdf5r::H5RefClass$get_file_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_obj_type">}\href{../../hdf5r/html/H5RefClass.html#method-get_obj_type}{\code{hdf5r::H5RefClass$get_obj_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="get_ref">}\href{../../hdf5r/html/H5RefClass.html#method-get_ref}{\code{hdf5r::H5RefClass$get_ref()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="inc_ref">}\href{../../hdf5r/html/H5RefClass.html#method-inc_ref}{\code{hdf5r::H5RefClass$inc_ref()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5RefClass" data-id="methods">}\href{../../hdf5r/html/H5RefClass.html#method-methods}{\code{hdf5r::H5RefClass$methods()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete">}\href{../../hdf5r/html/H5File.html#method-attr_delete}{\code{hdf5r::H5File$attr_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete_by_idx">}\href{../../hdf5r/html/H5File.html#method-attr_delete_by_idx}{\code{hdf5r::H5File$attr_delete_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_delete_by_name">}\href{../../hdf5r/html/H5File.html#method-attr_delete_by_name}{\code{hdf5r::H5File$attr_delete_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_exists">}\href{../../hdf5r/html/H5File.html#method-attr_exists}{\code{hdf5r::H5File$attr_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_exists_by_name">}\href{../../hdf5r/html/H5File.html#method-attr_exists_by_name}{\code{hdf5r::H5File$attr_exists_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_get_number">}\href{../../hdf5r/html/H5File.html#method-attr_get_number}{\code{hdf5r::H5File$attr_get_number()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_info_by_idx">}\href{../../hdf5r/html/H5File.html#method-attr_info_by_idx}{\code{hdf5r::H5File$attr_info_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_info_by_name">}\href{../../hdf5r/html/H5File.html#method-attr_info_by_name}{\code{hdf5r::H5File$attr_info_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_name_by_idx">}\href{../../hdf5r/html/H5File.html#method-attr_name_by_idx}{\code{hdf5r::H5File$attr_name_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open">}\href{../../hdf5r/html/H5File.html#method-attr_open}{\code{hdf5r::H5File$attr_open()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open_by_idx">}\href{../../hdf5r/html/H5File.html#method-attr_open_by_idx}{\code{hdf5r::H5File$attr_open_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_open_by_name">}\href{../../hdf5r/html/H5File.html#method-attr_open_by_name}{\code{hdf5r::H5File$attr_open_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_rename">}\href{../../hdf5r/html/H5File.html#method-attr_rename}{\code{hdf5r::H5File$attr_rename()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="attr_rename_by_name">}\href{../../hdf5r/html/H5File.html#method-attr_rename_by_name}{\code{hdf5r::H5File$attr_rename_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="close_all">}\href{../../hdf5r/html/H5File.html#method-close_all}{\code{hdf5r::H5File$close_all()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="commit">}\href{../../hdf5r/html/H5File.html#method-commit}{\code{hdf5r::H5File$commit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_attr">}\href{../../hdf5r/html/H5File.html#method-create_attr}{\code{hdf5r::H5File$create_attr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_attr_by_name">}\href{../../hdf5r/html/H5File.html#method-create_attr_by_name}{\code{hdf5r::H5File$create_attr_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_dataset">}\href{../../hdf5r/html/H5File.html#method-create_dataset}{\code{hdf5r::H5File$create_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_group">}\href{../../hdf5r/html/H5File.html#method-create_group}{\code{hdf5r::H5File$create_group()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="create_reference">}\href{../../hdf5r/html/H5File.html#method-create_reference}{\code{hdf5r::H5File$create_reference()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="exists">}\href{../../hdf5r/html/H5File.html#method-exists}{\code{hdf5r::H5File$exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="file_info">}\href{../../hdf5r/html/H5File.html#method-file_info}{\code{hdf5r::H5File$file_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="flush">}\href{../../hdf5r/html/H5File.html#method-flush}{\code{hdf5r::H5File$flush()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_filename">}\href{../../hdf5r/html/H5File.html#method-get_filename}{\code{hdf5r::H5File$get_filename()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_filesize">}\href{../../hdf5r/html/H5File.html#method-get_filesize}{\code{hdf5r::H5File$get_filesize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_intent">}\href{../../hdf5r/html/H5File.html#method-get_intent}{\code{hdf5r::H5File$get_intent()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_count">}\href{../../hdf5r/html/H5File.html#method-get_obj_count}{\code{hdf5r::H5File$get_obj_count()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_ids">}\href{../../hdf5r/html/H5File.html#method-get_obj_ids}{\code{hdf5r::H5File$get_obj_ids()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="get_obj_name">}\href{../../hdf5r/html/H5File.html#method-get_obj_name}{\code{hdf5r::H5File$get_obj_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info">}\href{../../hdf5r/html/H5File.html#method-group_info}{\code{hdf5r::H5File$group_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info_by_idx">}\href{../../hdf5r/html/H5File.html#method-group_info_by_idx}{\code{hdf5r::H5File$group_info_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="group_info_by_name">}\href{../../hdf5r/html/H5File.html#method-group_info_by_name}{\code{hdf5r::H5File$group_info_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link">}\href{../../hdf5r/html/H5File.html#method-link}{\code{hdf5r::H5File$link()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_copy_from">}\href{../../hdf5r/html/H5File.html#method-link_copy_from}{\code{hdf5r::H5File$link_copy_from()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_copy_to">}\href{../../hdf5r/html/H5File.html#method-link_copy_to}{\code{hdf5r::H5File$link_copy_to()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_external">}\href{../../hdf5r/html/H5File.html#method-link_create_external}{\code{hdf5r::H5File$link_create_external()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_hard">}\href{../../hdf5r/html/H5File.html#method-link_create_hard}{\code{hdf5r::H5File$link_create_hard()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_create_soft">}\href{../../hdf5r/html/H5File.html#method-link_create_soft}{\code{hdf5r::H5File$link_create_soft()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_delete">}\href{../../hdf5r/html/H5File.html#method-link_delete}{\code{hdf5r::H5File$link_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_delete_by_idx">}\href{../../hdf5r/html/H5File.html#method-link_delete_by_idx}{\code{hdf5r::H5File$link_delete_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_exists">}\href{../../hdf5r/html/H5File.html#method-link_exists}{\code{hdf5r::H5File$link_exists()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_info">}\href{../../hdf5r/html/H5File.html#method-link_info}{\code{hdf5r::H5File$link_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_info_by_idx">}\href{../../hdf5r/html/H5File.html#method-link_info_by_idx}{\code{hdf5r::H5File$link_info_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_move_from">}\href{../../hdf5r/html/H5File.html#method-link_move_from}{\code{hdf5r::H5File$link_move_from()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_move_to">}\href{../../hdf5r/html/H5File.html#method-link_move_to}{\code{hdf5r::H5File$link_move_to()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_name_by_idx">}\href{../../hdf5r/html/H5File.html#method-link_name_by_idx}{\code{hdf5r::H5File$link_name_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_value">}\href{../../hdf5r/html/H5File.html#method-link_value}{\code{hdf5r::H5File$link_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="link_value_by_idx">}\href{../../hdf5r/html/H5File.html#method-link_value_by_idx}{\code{hdf5r::H5File$link_value_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="ls">}\href{../../hdf5r/html/H5File.html#method-ls}{\code{hdf5r::H5File$ls()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="mount">}\href{../../hdf5r/html/H5File.html#method-mount}{\code{hdf5r::H5File$mount()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_copy_from">}\href{../../hdf5r/html/H5File.html#method-obj_copy_from}{\code{hdf5r::H5File$obj_copy_from()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_copy_to">}\href{../../hdf5r/html/H5File.html#method-obj_copy_to}{\code{hdf5r::H5File$obj_copy_to()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info">}\href{../../hdf5r/html/H5File.html#method-obj_info}{\code{hdf5r::H5File$obj_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info_by_idx">}\href{../../hdf5r/html/H5File.html#method-obj_info_by_idx}{\code{hdf5r::H5File$obj_info_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="obj_info_by_name">}\href{../../hdf5r/html/H5File.html#method-obj_info_by_name}{\code{hdf5r::H5File$obj_info_by_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="open">}\href{../../hdf5r/html/H5File.html#method-open}{\code{hdf5r::H5File$open()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="open_by_idx">}\href{../../hdf5r/html/H5File.html#method-open_by_idx}{\code{hdf5r::H5File$open_by_idx()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="path_valid">}\href{../../hdf5r/html/H5File.html#method-path_valid}{\code{hdf5r::H5File$path_valid()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="print">}\href{../../hdf5r/html/H5File.html#method-print}{\code{hdf5r::H5File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hdf5r" data-topic="H5File" data-id="unmount">}\href{../../hdf5r/html/H5File.html#method-unmount}{\code{hdf5r::H5File$unmount()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{scdisk} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scdisk$new(
  filename = NULL,
  mode = c("a", "r", "r+", "w", "w-", "x"),
  validate = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Name of on-disk file to connect to}

\item{\code{mode}}{How to open the file, choose from:
\describe{
 \item{a}{Create new or open existing file, allow read and write}
 \item{r}{Open existing file, allow read only}
 \item{r+}{Open existing file, allow read and write}
 \item{w}{Create new file (deleting any existing one), allow read and write}
 \item{w-, x}{Create new file (error if exists), allow read and write}
}}

\item{\code{validate}}{Validate the file upon connection}

\item{\code{...}}{Extra arguments passed to validation routine}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalizer"></a>}}
\if{latex}{\out{\hypertarget{method-finalizer}{}}}
\subsection{Method \code{finalizer()}}{
Handle the loss of reference to this \code{scdisk} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scdisk$finalizer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chunk.points"></a>}}
\if{latex}{\out{\hypertarget{method-chunk.points}{}}}
\subsection{Method \code{chunk.points()}}{
Generate chunk points for a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scdisk$chunk.points(dataset, MARGIN = NULL, csize = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{Name of dataset}

\item{\code{MARGIN}}{Direction to chunk in; defaults to last dimension of dataset}

\item{\code{csize}}{Size of chunk; defaults to hdf5r-suggested chunk size}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix where each row is a chunk, column 1 is start points,
column 2 is end points
}
}
}
