% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratDisk-package}
\alias{SeuratDisk-package}
\alias{_PACKAGE}
\alias{SeuratDisk}
\title{SeuratDisk: Interfaces for HDF5-Based Single Cell File Formats}
\description{
As single-cell expression datasets grow in size, the computational
  requirements to store and handle them grow as well. To address the issues posed
  by large datasets, SeuratDisk provides an interface to various single-cell
  oriented HDF5-based file formats and provides infrastructure to perform
  disk-based computation on said files. In addition, SeuratDisk integrates with
  the Seurat framework and provides disk-based methods for Seurat's toolkit.
}
\section{Package options}{


SeuratDisk uses the following options to control behaviour, users can configure
these with \code{\link[base]{options}}:

\describe{
 \item{\code{SeuratDisk.dtypes.logical_to_int}}{When writing
 \link[base]{logical} vectors, coerce to integer types to ensure
 compatibility across languages (see \code{\link{BoolToInt}} for more details)}
 \item{\code{SeuratDisk.dtypes.dataframe_as_group}}{When writing
 \link[base]{data.frame}s, always write out as a group regardless of factor
 presence}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mojaveazure.github.io/seurat-disk/}
  \item \url{https://github.com/mojaveazure/seurat-disk}
  \item Report bugs at \url{https://github.com/mojaveazure/seurat-disk/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Other contributors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID}) [contributor]
}

}
