% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratDisk-pacakge}
\alias{SeuratDisk}
\alias{SeuratDisk-package}
\title{SeuratDisk: Interfaces for HDF5-Based Single Cell File Formats}
\description{
As single-cell expression datasets grow in size, the computational
  requirements to store and handle them grow as well. To address the issues posed
  by large datasets, SeuratDisk provides an interface to various single-cell
  oriented HDF5-based file formats and provides infrastructure to perform
  disk-based computation on said files. In addition, SeuratDisk integrates with
  the Seurat framework and provides disk-based methods for Seurat's toolkit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojaveazure/seurat-disk}
  \item Report bugs at \url{https://github.com/mojaveazure/seurat-disk/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Other contributors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID}) [contributor]
}

}
