% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratDisk-package}
\alias{SeuratDisk-package}
\alias{_PACKAGE}
\alias{SeuratDisk}
\title{SeuratDisk: Interfaces for HDF5-Based Single Cell File Formats}
\description{
The h5Seurat file format is specifically designed for the storage
  and analysis of multi-modal single-cell and spatially-resolved expression
  experiments, for example, from CITE-seq or 10X Visium technologies. It holds
  all molecular information and associated metadata, including (for example)
  nearest-neighbor graphs, dimensional reduction information, spatial
  coordinates and image data, and cluster labels. We also support rapid and
  on-disk conversion between h5Seurat and AnnData objects, with the goal of
  enhancing interoperability between Seurat and Scanpy.
}
\section{Package options}{


SeuratDisk uses the following options to control behaviour, users can configure
these with \code{\link[base]{options}}:

\describe{
 \item{\code{SeuratDisk.dtypes.logical_to_int}}{
  When writing \link[base]{logical} vectors, coerce to integer types to
  ensure compatibility across languages (see \code{\link{BoolToInt}} for
  more details)
 }
 \item{\code{SeuratDisk.dtypes.dataframe_as_group}}{
  When writing \link[base]{data.frame}s, always write out as a group
  regardless of factor presence
 }
 \item{\code{SeuratDisk.chunking.MARGIN}}{
  Default direction for chunking datasets; choose from:
  \describe{
   \item{largest}{Chunk along the largest dimension of a dataset}
   \item{smallest}{Chunk along the smallest dimension}
   \item{first}{Chunk along the first dimension}
   \item{last}{Chunk along the last dimension}
  }
 }
 \item{\code{SeuratDisk.dimreducs.allglobal}}{
  Treat all DimReducs as global, regardless of actual global status
 }
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mojaveazure.github.io/seurat-disk/}
  \item \url{https://github.com/mojaveazure/seurat-disk}
  \item Report bugs at \url{https://github.com/mojaveazure/seurat-disk/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Other contributors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID}) [contributor]
}

}
