% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteH5Group.R
\name{WriteH5Group}
\alias{WriteH5Group}
\title{Write data to an HDF5 group}
\usage{
WriteH5Group(x, name, hgroup, verbose = TRUE)
}
\arguments{
\item{x}{An object}

\item{name}{Name to save data as}

\item{verbose}{Show progress updates}

\item{h5group}{An HDF5 file or group (\code{H5File} or \code{H5Group} objects
from hdf5r)}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Write data to an HDF5 group
}
